requirements.txt
fastapi
uvicorn
python-dotenv
boto3
botocore
b2sdk
httpx
sqlalchemy
alembic
psycopg2-binary
pandas
numpy
scikit-learn
